requirements.txt
# Core Dependencies (Cross-platform)
pynput==1.7.6
psutil==5.9.6

# Platform-specific window tracking
# Windows: pywin32
pywin32==306; sys_platform == 'win32'

# Linux: python-xlib (for X11 window tracking)
python-xlib==0.33; sys_platform == 'linux'

# macOS: pyobjc frameworks
pyobjc-framework-Cocoa==10.1; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.1; sys_platform == 'darwin'

# Data Processing (Cross-platform)
pandas==2.1.4
numpy==1.24.3

# Machine Learning (Cross-platform)
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Visualization (Cross-platform)
matplotlib==3.8.2
seaborn==0.13.0

# Utilities (Cross-platform)
python-dateutil==2.8.2
tqdm==4.66.1

# Visual Intelligence (v2.1 Enhancement - Cross-platform)
mss==9.0.1  # Screenshot capture
Pillow==10.1.0  # Image processing
opencv-python==4.8.1.78  # Computer vision
pytesseract==0.3.10  # OCR text extraction
scikit-image==0.22.0  # SSIM and image metrics

# Development (Cross-platform)
pytest==7.4.3
pytest-cov==4.1.0
